Move shared view models from Document to Content
