+++ src/Umbraco.Cms.Api.Management/ViewModels/Document/PublishDocumentRequestModel.cs
@@ -1,3 +1,5 @@
+using Umbraco.Cms.Api.Management.ViewModels.Content;
+
 namespace Umbraco.Cms.Api.Management.ViewModels.Document;
 
 /// <summary>
@@ -9,35 +11,4 @@ public class PublishDocumentRequestModel
     /// Gets or sets the collection of publish schedules for different cultures.
     /// </summary>
     public required IEnumerable<CultureAndScheduleRequestModel> PublishSchedules { get; set; }
-}
-
-// TODO ELEMENTS: move the following classes to ViewModels.Content
-/// <summary>
-/// Request model for specifying culture and scheduling settings for a document.
-/// </summary>
-public class CultureAndScheduleRequestModel
-{
-    /// <summary>
-    /// Gets or sets the culture. Null means invariant.
-    /// </summary>
-    public string? Culture { get; set; }
-
-    /// <summary>
-    /// Gets or sets the schedule of publishing. Null means immediately.
-    /// </summary>
-    public ScheduleRequestModel? Schedule { get; set; }
-}
-
-/// <summary>
-/// Request model for scheduling the publication of a document.
-/// </summary>
-public class ScheduleRequestModel
-{
-    /// <summary>Gets or sets the scheduled publish time for the document.</summary>
-    public DateTimeOffset? PublishTime { get; set; }
-
-    /// <summary>
-    /// Gets or sets the time when the document should be unpublished.
-    /// </summary>
-    public DateTimeOffset? UnpublishTime { get; set; }
 }
